refactor: architecture domains
